add functionality on checkbox
