+++ src/index.js
@@ -1,4 +1,5 @@
 import './style.css';
+import displayTask from './modules/displayTask.js';
 
 const form = document.getElementById('form');
 const task = document.getElementById('task');
@@ -8,42 +9,6 @@ let listName = [];
 
 const initList = () => {
   listName = localStorage.getItem('tasks') ? JSON.parse(localStorage.getItem('tasks')) : [];
-};
-
-// make task editable
-const makeTaskEditable = () => {
-  const taskAreas = document.querySelectorAll('.area');
-  taskAreas.forEach((taskArea, index) => {
-    taskArea.addEventListener('blur', () => {
-      listName[index].text = taskArea.innerText;
-      localStorage.setItem('tasks', JSON.stringify(listName));
-    });
-    taskArea.addEventListener('keydown', (e) => {
-      if (e.keyCode === 13) {
-        e.preventDefault();
-        listName[index].text = taskArea.innerText;
-        localStorage.setItem('tasks', JSON.stringify(listName));
-        taskArea.blur();
-      }
-    });
-  });
-};
-
-// display task
-const displayTask = () => {
-  items.innerHTML = '';
-  for (let i = 0; i < listName.length; i += 1) {
-    const isChecked = listName[i].completed ? 'checked' : '';
-    items.innerHTML += `
-      <div class="to-do-item">
-        <input id="check${i}" type="checkbox" ${isChecked}>
-        <div id="area${i}" class="area" contentEditable="true">${listName[i].text}</div>
-        <button id="delete${i}" class="deleteBtn"><i class="fa-solid fa-trash-can"></i></button>
-      </div>
-    `;
-    task.value = '';
-  }
-  makeTaskEditable();
 };
 
 // adding task
@@ -82,3 +47,17 @@ window.onload = () => {
   initList();
   displayTask();
 };
+
+const clear = document.getElementById('clear');
+
+clear.addEventListener('click', () => {
+  let todo = JSON.parse(localStorage.getItem('tasks'));
+  todo = todo.filter((item) => !item.completed);
+  localStorage.setItem('tasks', JSON.stringify(todo));
+  for (let i = 0; i < todo.length; i += 1) {
+    const todo = JSON.parse(localStorage.getItem('tasks')) || [];
+    todo[i].index = i + 1;
+    localStorage.setItem('tasks', JSON.stringify(todo));
+  }
+  displayTask();
+});
